+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.1.0
+
+nieuw:
+
+- Een kans om griekse letters te veranderen in latijnse in de config (0-100)
+- Support voor `config.gebruikAlternatieveVragenAfwisselendWanneerBeschikbaar`
+- currentItem om de huidige vraag op te halen
+
+notes:
+
+- Je MOET de currentItem syntax gebruiken om de vraag op te halen anders werken de vraag gerelateerde configs niet
+
 # 1.0.5
 
 - support voor optionele delen van antwoorden zonder haakjes, bijv `antwoord (optioneel deel)` kan ook beantwoord worden met `antwoord optioneel deel`